move utils functions from `ManagedStrategy` to `utils` subpackage
